create constants folder for queries
