+++ client/src/FullProduct.css
@@ -13,7 +13,7 @@
   margin: 5% auto;
   max-width: 820px;
   min-width: 700px;
-  min-height: 450px;
+  min-height: 520px;
   padding: 40px 30px;
   background: #fff;
 }
