+++ client/src/components/Products.jsx
@@ -1,66 +1,26 @@
 import React, { useState } from 'react';
-import { useQuery, gql } from '@apollo/client';
+import { useQuery } from '@apollo/client';
 import Product from './Product';
 import Categories from './Categories';
-
-const GET_DATA = gql`
-  query myQuery($categoryInput: CategoryInput) {
-    category(input: $categoryInput) {
-      name
-      products {
-        id
-        name
-        inStock
-        gallery
-        description
-        category
-        attributes {
-          name
-          id
-          type
-          items {
-            id
-            value
-            displayValue
-          }
-        }
-        prices {
-          amount
-          currency {
-            label
-            symbol
-          }
-        }
-        brand
-      }
-    }
-  }
-`;
-const GET_CATEGORIES = gql`
-  query myAwesomeQuery {
-    categories {
-      name
-    }
-  }
-`;
+import { QUERIES } from '../constants/costants';
 
 const Products = () => {
   const [, setCategoryInput] = useState('all');
-  const { error, data, loading, refetch: refetchProducts } = useQuery(GET_DATA);
+  const { error, data, loading, refetch: refetchProducts } = useQuery(QUERIES.GET_DATA);
   const {
     error: errorCategories,
     data: dataCategories,
     loading: loadingCategories,
-  } = useQuery(GET_CATEGORIES);
-
-  console.log('Data', { error, loading, data, errorCategories, loadingCategories });
+  } = useQuery(QUERIES.GET_CATEGORIES);
 
   return (
     <div>
       <div className="container my-3 py-1">
         <div className="row justify-content-center">
           <div className="col-12 mb-5">
-            {data ? (
+            {loading ? (
+              <>Loading...</>
+            ) : (
               <>
                 <Categories
                   allCategories={dataCategories?.categories}
@@ -73,8 +33,6 @@ const Products = () => {
                   ))}
                 </div>
               </>
-            ) : (
-              <>Loading...</>
             )}
           </div>
         </div>
